Represent timestamps as milliseconds in a long in Java-land.
When serializing to JSON, we still keep the timestamp in seconds with a
fracitonal part. That leaves open the possibility of another platform
having higher precision than milliseconds.

Fixed #47.

+++ openxc/src/com/openxc/measurements/BaseMeasurement.java
@@ -71,15 +71,15 @@ public class BaseMeasurement<TheUnit extends Unit> implements Measurement {
         mRange = range;
     }
 
-    public void setTimestamp(double timestamp) {
+    public void setTimestamp(long timestamp) {
         mValue.setTimestamp(timestamp);
     }
 
-    public double getAge() {
+    public long getAge() {
         return mValue.getAge();
     }
 
-    public double getBirthtime() {
+    public long getBirthtime() {
         return mValue.getTimestamp();
     }
